% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcosso.cox.R
\name{cdcosso.cox}
\alias{cdcosso.cox}
\title{Load a matrix from a file}
\usage{
cdcosso.cox(
  x,
  time,
  status,
  wt,
  lambda0,
  lambda_theta,
  gamma,
  type,
  kparam,
  scale,
  algo
)
}
\arguments{
\item{x}{Explanation variable matrix or data frame.}

\item{time}{Dependent variable vector or matrix or data frame containing time and status columns (for Cox model).}

\item{status}{Type of statistical model. Use one of the following strings: "gaussian", "binomial", "poisson", "negbin", "svm", or "Cox".}

\item{wt}{Type of statistical model. Use one of the following strings: "gaussian", "binomial", "poisson", "negbin", "svm", or "Cox".}

\item{lambda0}{Type of kernel function to use in case of SVM model. Use one of the following strings: "linear", "gaussian", "poly", "spline", "anova_gaussian", or "gaussian2".}

\item{lambda_theta}{Type of optimization algorithm. Use either the string "CD" (Coordinate Descent) or "QP".}

\item{gamma}{Kernel parameter values to use for SVM models.}

\item{type}{Gamma value used in Stochastic Search Optimization.}

\item{kparam}{Number of folds for cross-validation.}

\item{scale}{Boolean for whether to scale the input data to range between 0 and 1.}

\item{algo}{Logical value indicating whether to standardize explanatory variables.}
}
\value{
A list containing information about the fitted model. Depending on the type of dependent variable, various information may be returned.
}
\description{
The cdcosso function is a function that solves component selection using the Coordinate Descent algorithm.
This function can be applied to various response variables, continuous, count, binary, and survival analysis.
Because it is a type of nonparametric inference, various types of kernels can be selected.
To select hyperparameters, the function is designed to perform cross-validation.
}
