% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cossonet.cox.R
\name{cossonet.cox}
\alias{cossonet.cox}
\title{The cossonet.cox function is a COSSO model based on the Cox proportional hazards model applied to survival responses.
Survival responses should be in matrix form with time and status as column names. It works internally via the cossonet function.}
\usage{
cossonet.cox(
  x,
  time,
  status,
  nbasis,
  basis.id,
  wt,
  lambda0,
  lambda_theta,
  gamma,
  type,
  nfold,
  kparam,
  one.std,
  scale
)
}
\arguments{
\item{x}{Input matrix or data frame of $n$ by $p$. \code{x} must have at least two columns ($p>1$).}

\item{time}{for right censored data, this is the follow up time. For interval data, the first argument is the starting time for the interval. It follows the same input format as the \code{time} argument in the \code{Surv} function from the \code{survival} package.}

\item{status}{Status indicator, typically coded as 0 for "alive" and 1 for "dead." It follows the same input format as the \code{status} argument in the \code{Surv} function from the \code{survival} package.}

\item{nbasis}{The number of "knots". If \code{basis.id} is provided, it is set to the length of \code{basis.id}.}

\item{basis.id}{The index of the "knot" to select.}

\item{wt}{The weights assigned to the explanatory variables.}

\item{lambda0}{A vector of \code{lambda0} sequences.}

\item{lambda_theta}{A vector of \code{lambda} sequences.}

\item{gamma}{Elastic-net mixing parameter.}

\item{type}{The kernel function.}

\item{nfold}{The number of folds to use in cross-validation is used to determine how many subsets to divide the data into for the training and validation sets.}

\item{kparam}{Parameters for Gaussian and polynomial kernel functions.}

\item{one.std}{A logical value indicating whether to apply the "1-standard error rule," selecting the simplest model within one standard error of the best model.}

\item{scale}{Boolean for whether to scale continuous explanatory variables to values between 0 and 1.}
}
\value{
A list of outputs obtained from the fitted model for the Cox proportional hazards model
}
\description{
The cossonet.cox function is a COSSO model based on the Cox proportional hazards model applied to survival responses.
Survival responses should be in matrix form with time and status as column names. It works internally via the cossonet function.
}
