% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cossonet.exp.R
\name{cossonet.exp}
\alias{cossonet.exp}
\title{The cossonet.exp function is a COSSO model applied to responses belonging to the exponential family.
This function is applied to continuous, binary, and count responses.
It operates internally through the cossonet function.}
\usage{
cossonet.exp(
  x,
  y,
  wt,
  nbasis,
  basis.id,
  lambda0,
  lambda_theta,
  gamma,
  obj,
  type,
  nfold,
  kparam,
  one.std,
  scale
)
}
\arguments{
\item{x}{Input matrix or data frame of $n$ by $p$. \code{x} must have at least two columns ($p>1$).}

\item{y}{A response vector with a continuous, binary, or count type.}

\item{wt}{The weights assigned to the explanatory variables.}

\item{nbasis}{The number of "knots". If \code{basis.id} is provided, it is set to the length of \code{basis.id}.}

\item{basis.id}{The index of the "knot" to select.}

\item{lambda0}{A vector of \code{lambda0} sequences.}

\item{lambda_theta}{A vector of \code{lambda} sequences.}

\item{gamma}{Elastic-net mixing parameter.}

\item{obj}{distribution used in the model.}

\item{type}{The kernel function. One of four types of \code{linear} (default), \code{gaussian}, \code{poly}, and \code{spline}.}

\item{nfold}{The number of folds to use in cross-validation is used to determine how many subsets to divide the data into for the training and validation sets.}

\item{kparam}{Parameters for Gaussian and polynomial kernel functions.}

\item{one.std}{A logical value indicating whether to apply the "1-standard error rule," selecting the simplest model within one standard error of the best model.}

\item{scale}{Boolean for whether to scale continuous explanatory variables to values between 0 and 1. Default of \code{TRUE}.}
}
\value{
A list of outputs obtained from a model fitted to an exponential distribution.
}
\description{
The cossonet.exp function is a COSSO model applied to responses belonging to the exponential family.
This function is applied to continuous, binary, and count responses.
It operates internally through the cossonet function.
}
