% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cossonet.predict.R
\name{cossonet.predict}
\alias{cossonet.predict}
\title{Load a matrix from a file}
\usage{
cossonet.predict(model, testx)
}
\arguments{
\item{model}{Fitted model from cdcosso.}

\item{testx}{A new dataset to predict.}
}
\value{
A list containing the predicted value for the new dataset.
}
\description{
The predict.cdcosso function is a function that receives the object and test data (testx) of the cdcosso package as input
and generates a predicted value for the test data.
This function uses the given test data to calculate predictions from the weights and biases generated by the model.
}
