% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cossonet.R
\name{cossonet}
\alias{cossonet}
\title{Load a matrix from a file}
\usage{
cossonet(
  x,
  y,
  family = c("gaussian", "binomial", "poisson", "Cox"),
  wt = rep(1, ncol(x)),
  scale = TRUE,
  nbasis,
  basis.id,
  kernel = c("linear", "gaussian", "poly", "spline"),
  effect = c("main", "interaction"),
  kparam = 1,
  lambda0 = exp(seq(log(2^{
     -10
 }), log(2^{
     10
 }), length.out = 20)),
  lambda_theta = exp(seq(log(2^{
     -10
 }), log(2^{
     10
 }), length.out = 20)),
  gamma = 0.95
)
}
\arguments{
\item{x}{Input matrix of $n$ by $p$, where each row is an observation. A matrix or dataframe. x must have at least two columns ($p>1$).}

\item{y}{The response variable. If family="gaussian" or family="poisson" (non-negative counts), it is quantitative. If family="binomial", it must be a vector with two levels. If family="cox", y must be a two-column matrix (or dataframe) with columns named 'time' and 'state'.}

\item{family}{The response type}

\item{wt}{The weights of the predictors. The default is \code{rep(1,ncol(x))}.}

\item{scale}{If TRUE, continuous predictors are rescaled to the interval \verb{[0,1]}. The default is \code{TRUE}.}

\item{nbasis}{The number of "knots" to choose from. If basis.id is provided, it is ignored.}

\item{basis.id}{An index that specifies the selected "knot".}

\item{kernel}{The kernel function. Four types are provided: \code{linear} (default), \code{gaussian}, \code{poly}, \code{spline}.}

\item{effect}{The effect of the component. \code{main} (default) for the main effect, \code{interaction} for interactions.}

\item{kparam}{Parameters for the kernel function. Used by Gaussian and polynomial kernels.}

\item{lambda0}{A vector of lambda0 sequences. The default is \code{exp()}. This may need to be adjusted based on your data. Do not provide a single value for lambda0.}

\item{lambda_theta}{A vector of lambda_theta sequences. The default is \code{exp()}. This may need to be adjusted based on your data. Do not provide a single value for lambda_theta.}

\item{gamma}{Elastic mesh mixing parameter, \verb{0 \\leq \\gamma \\leq 1}. When \code{gamma = 1} it uses LASSO penalty, when \code{gamma = 0} it uses Ridge penalty. The default is \code{gamma = 0.95}.}
}
\value{
A list containing information about the fitted model.
}
\description{
The cdcosso function is a function that solves component selection using the Coordinate Descent algorithm.
This function can be applied to various response variables, continuous, count, binary, and survival analysis.
Because it is a type of nonparametric inference, various types of kernels can be selected.
To select hyperparameters, the function is designed to perform cross-validation.
}
