% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLD.R
\name{KLD}
\alias{KLD}
\title{Compute Kullback-Leibler Divergence}
\usage{
KLD(y, fhat)
}
\arguments{
\item{y}{A vector of true response variable.}

\item{fhat}{A vector of estimated response variable fitted by \code{cdcosso}.}
}
\value{
A vector of Kullback-Leibler divergence evaluated by test data.
}
\description{
Compute Kullback-Leibler Divergence
}
