% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric}
\alias{metric}
\title{The function \code{metric} provides a contingency table for the predicted class and the true class for binary classes.}
\usage{
metric(true, est)
}
\arguments{
\item{true}{binary true class.}

\item{est}{binary predicted class.}
}
\value{
a contingency table for the predicted results of binary class responses.
}
\description{
The function \code{metric} provides a contingency table for the predicted class and the true class for binary classes.
}
