% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partial_Lik.R
\name{Partial_Lik}
\alias{Partial_Lik}
\title{Load a matrix from a file}
\usage{
Partial_Lik(time, status, K, a)
}
\arguments{
\item{time}{A vector of obbseved time.}

\item{status}{A vector of observed event indicator.}

\item{K}{Kernel function evaluated in every component.}

\item{a}{A vector evaluating observation.}
}
\value{
Partial log-likelihood of Cox proportional model.
}
\description{
The cdcosso function is a function that solves component selection using the Coordinate Descent algorithm.
This function can be applied to various response variables, continuous, count, binary, and survival analysis.
Because it is a type of nonparametric inference, various types of kernels can be selected.
To select hyperparameters, the function is designed to perform cross-validation.
}
