% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcosso.R
\name{cdcosso}
\alias{cdcosso}
\title{Load a matrix from a file}
\usage{
cdcosso(
  x,
  y,
  family = c("gaussian", "binomial", "poisson", "Cox"),
  kernel = c("linear", "gaussian", "poly", "spline"),
  effect = c("main", "interaction"),
  algo = c("CD", "QP"),
  kparam = 1,
  lambda0 = exp(seq(log(2^{
     -6
 }), log(2^{
     6
 }), length.out = 20)),
  lambda_theta = exp(seq(log(2^{
     -10
 }), log(2^{
     4
 }), length.out = 20)),
  gamma = 0.95,
  scale = TRUE
)
}
\arguments{
\item{x}{Input matrix with $n$ observations and $p$ dimension.}

\item{y}{Response variable. The type can be continuous (default), binary class, non-negative count, survival.}

\item{family}{A character string representing one of the built-in families. The value depends on the type of response variable, \code{family='gaussian'} for continuous, \code{family='binomial'} forj binary class, \code{family='poisson'} for non-negative count , and \code{family='Cox'} for survival.}

\item{kernel}{Kernel function which is used to convert the input data for training and predicting. The four types is provided, \code{linear} (default), \code{gaussian}, \code{poly}, and \code{spline}.}

\item{effect}{Effect of the component to be analyzed, \code{effect = "main"} for main effect (default), and \code{effect = "interaction"} for interaction.}

\item{algo}{Type of optimization algorithm. Use either the string "CD" (Coordinate Descent) or "QP".}

\item{kparam}{Kernel parameter values that is used in gaussian kernel and polynomial kernel.}

\item{lambda0}{A vector of tuning parameter to select optimal smoothing parameter.}

\item{lambda_theta}{A vector of tuning parameter to select optimal tuning parameter.}

\item{gamma}{The elastic net mixing parameter between 0 and 1. When \code{gamma = 1}, the penalty is to be LASSO. When \code{gamma = 0}, the penalty is to be ridge penalty. The default is \code{gamma = 0.95}.}

\item{scale}{Boolean for whether to scale the input data to range between 0 and 1.}
}
\value{
A list containing information about the fitted model.
}
\description{
The cdcosso function is a function that solves component selection using the Coordinate Descent algorithm.
This function can be applied to various response variables, continuous, count, binary, and survival analysis.
Because it is a type of nonparametric inference, various types of kernels can be selected.
To select hyperparameters, the function is designed to perform cross-validation.
}
