% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cossonet.exp.R
\name{cossonet.exp}
\alias{cossonet.exp}
\title{Load a matrix from a file}
\usage{
cossonet.exp(
  x,
  y,
  f,
  cv,
  wt,
  nbasis,
  basis.id,
  lambda0,
  lambda_theta,
  gamma,
  obj,
  type,
  kparam,
  scale
)
}
\arguments{
\item{x}{Input matrix with $n$ observations and $p$ dimension.}

\item{y}{Response variable. The type can be continuous (default), binary class, non-negative count, survival.}

\item{f}{The true function.}

\item{cv}{measure for cross-validation}

\item{wt}{A weight vector for components.}

\item{nbasis}{The number of basis.}

\item{basis.id}{The index of basis.}

\item{lambda0}{A vector of tuning parameter to select optimal smoothing parameter.}

\item{lambda_theta}{A vector of tuning parameter to select optimal tuning parameter.}

\item{gamma}{The elastic net mixing parameter between 0 and 1. When \code{gamma = 1}, the penalty is to be LASSO. When \code{gamma = 0}, the penalty is to be ridge penalty. The default is \code{gamma = 0.95}.}

\item{obj}{The type of family.}

\item{type}{Kernel function which is used to convert the input data for training and predicting. The four types is provided, \code{linear} (default), \code{gaussian}, \code{poly}, and \code{spline}.}

\item{kparam}{Kernel parameter values that is used in gaussian kernel and polynomial kernel.}

\item{scale}{Boolean for whether to scale the input data to range between 0 and 1.}
}
\value{
A list containing information about the fitted model. Depending on the type of dependent variable, various information may be returned.
}
\description{
The cdcosso function is a function that solves component selection using the Coordinate Descent algorithm.
This function can be applied to various response variables, continuous, count, binary, and survival analysis.
Because it is a type of nonparametric inference, various types of kernels can be selected.
To select hyperparameters, the function is designed to perform cross-validation.
}
