% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cossonet.predict.R
\name{cossonet.predict}
\alias{cossonet.predict}
\title{The function \code{cossonet.predict} predicts predictive values for new data based on an object from the \code{cossonet} function.}
\usage{
cossonet.predict(model, testx)
}
\arguments{
\item{model}{The fitted cossonet object.}

\item{testx}{The new data set to be predicted.}
}
\value{
A list of predicted values for the new data set.
}
\description{
The function \code{cossonet.predict} predicts predictive values for new data based on an object from the \code{cossonet} function.
}
