% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric}
\alias{metric}
\title{Load a matrix from a file}
\usage{
metric(true, est)
}
\arguments{
\item{true}{A vector of true value of binary class response.}

\item{est}{A vector of estimated response by \code{cdcosso}.}
}
\value{
A list of contingency table for predicted results of binary class response.
}
\description{
The cdcosso function is a function that solves component selection using the Coordinate Descent algorithm.
This function can be applied to various response variables, continuous, count, binary, and survival analysis.
Because it is a type of nonparametric inference, various types of kernels can be selected.
To select hyperparameters, the function is designed to perform cross-validation.
}
