% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Partial_Lik.R
\name{Partial_Lik}
\alias{Partial_Lik}
\title{Load a matrix from a file}
\usage{
Partial_Lik(time, status, K, a)
}
\arguments{
\item{time}{A vector of true response variable.}

\item{status}{A vector of estimated response variable fitted by \code{cdcosso}.}

\item{K}{A family corresponding to the type of response variable.}

\item{a}{A family corresponding to the type of response variable.}
}
\value{
A vector of Kullback-Leibler divergence evaluated by test data.
}
\description{
The cdcosso function is a function that solves component selection using the Coordinate Descent algorithm.
This function can be applied to various response variables, continuous, count, binary, and survival analysis.
Because it is a type of nonparametric inference, various types of kernels can be selected.
To select hyperparameters, the function is designed to perform cross-validation.
}
